remove wayland server-side decorations support
After careful consideration, I decided to remove support for the xdg-decoration
protocol, which allowed us to ask the compositors for server-side decorations.

The reason for this is that the protocol was purposefully designed to be
impossible to use synchronously: the compositor's answer is communicated
through a callback without any way for us to know when it will happen,
or if it will happen at all since an answer isn't even required.

A partial solution would be to provide a user callback for execution in the
xdg-decoration callback, so we don't have to check the current status every
frame, but then it should be implemented for all other platforms as well.

Since nothing can portably explain the xdg-decoration callback being called,
it would be impossible to reproduce this behaviour on other platforms in
a way that is both logic and indicative of how it should be handled
for optimal portability.

I find it very sad that some people rushed to fill Wayland's "standard void"
with the assumption that the core protocol expects client-side decorations.
It is a bad thing in my humble opinion, and I would have loved supporting
the xdg-decoration protocol to make it easy to get server-side ones.

I think it is time to face the facts though: client-side decorations have
won and it is almost impossible to get server-side decorations reliably,
let alone in a way that allows abstraction libraries like Globox to
maintain portability across display systems.