do not link with libGLX and use libGL instead
Some Linux distos (like Alpine, which we use for CI) do not provide a
package-config information file for GLX. This kind of issue is usually
fixed by providing package-config with a custom `GLX.pc`, but such a
workaround is obviously not intended to be portable across distros.

Instead, we can fix this issue easily using some Linux nerd tricks.
Because of a standard weirdness, libGLs always reference libGLX.
Since libGLX obviously references libGL as well, it is possible
to link both interchangeably, as it will have the same effect.
libGL does not suffer from the same package-config void,
so we can use it safely in this case.